requirements.txt
asn1crypto==0.24.0
cryptography==2.1.4
enum34==1.1.6
fusepy==3.0.1
idna==2.6
ipaddress==1.0.17
keyring==10.6.0
keyrings.alt==3.0
pycrypto==2.6.1
pygobject==3.26.1
pyOpenSSL==17.5.0
pyxdg==0.25
SecretStorage==2.3.1
six==1.11.0
